test: testa produto e suas validacoes

+++ src/dominio/Produto/ProdutoEletronico.java
@@ -3,7 +3,7 @@ package dominio.Produto;
 import java.math.BigDecimal;
 
 public class ProdutoEletronico extends Produto {
-    public ProdutoEletronico(String nome, BigDecimal valorProduto) {
+    public ProdutoEletronico(String nome, double valorProduto) {
         super(nome, valorProduto, TipoProduto.ELETRONICO);
         if (!validarProdutoAtivo()) {
             throw new IllegalArgumentException("Produto eletrônico inválido.");
