+++ src/dominio/Produto/ProdutoLivro.java
@@ -3,7 +3,7 @@ package dominio.Produto;
 import java.math.BigDecimal;
 
 public class ProdutoLivro extends Produto {
-    public ProdutoLivro(String nome, BigDecimal valorProduto) {
+    public ProdutoLivro(String nome, double valorProduto) {
         super(nome, valorProduto, TipoProduto.LIVRO);
         if (!validarProdutoAtivo()) {
             throw new IllegalArgumentException("Produto livro inválido.");
