+++ src/dominio/Produto/ProdutoPapelaria.java
@@ -3,7 +3,7 @@ package dominio.Produto;
 import java.math.BigDecimal;
 
 public class ProdutoPapelaria extends Produto {
-    public ProdutoPapelaria(String nome, BigDecimal valorProduto) {
+    public ProdutoPapelaria(String nome, double valorProduto) {
         super(nome, valorProduto, TipoProduto.PAPELARIA);
         if (!validarProdutoAtivo()) {
             throw new IllegalArgumentException("Produto papelaria inválido.");
