+++ src/dominio/Produto/ProdutoVestuario.java
@@ -3,7 +3,7 @@ package dominio.Produto;
 import java.math.BigDecimal;
 
 public class ProdutoVestuario extends Produto {
-    public ProdutoVestuario(String nome, BigDecimal valorProduto) {
+    public ProdutoVestuario(String nome, double valorProduto) {
         super(nome, valorProduto, TipoProduto.VESTUARIO);
         if (!validarProdutoAtivo()) {
             throw new IllegalArgumentException("Produto vestuario inválido.");
